requirements.txt
# Packages
streamlit==1.28.0
numpy==1.26.1
pandas==2.1.1
plotly==5.17.0
xgboost==2.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.2
matplotlib==3.8.0

# Other dependencies
watchdog==2.1.6
st-annotated-text==4.0.1
faker==19.11.0

# Dev
black
isort
pytest==7.4.3